Develop rel 1 7 0

+++ src/iot_api.c
@@ -1252,7 +1252,7 @@ iot_error_t iot_misc_info_store(iot_misc_info_t type, const void *in_data)
 		json = JSON_PARSE(old_misc_info);
 		if (json == NULL) {
 			IOT_WARN("old misc_info(%s/%u) parsing failed",
-				old_misc_info, old_misc_info_len);
+				old_misc_info, (unsigned int)old_misc_info_len);
 			json = JSON_CREATE_OBJECT();
 		} else {
 			old_misc_avail = true;
